requirements.txt
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
idna==2.10
Pygments==2.7.3
requests==2.25.1
urllib3==1.26.2
mypy==0.812
pillow==8.3.2
cairosvg==2.5.2
